revert ArrayListComparatorFn int-> double
Issue:https://gitcode.com/openharmony/arkcompiler_runtime_core/issues/11628

+++ api/@ohos.util.ArrayList.d.ts
@@ -394,15 +394,14 @@ declare class ArrayList<T> {
  *
  * @param { T } firstValue - firstValue (required) previous element.
  * @param { T } secondValue - secondValue (required) next element.
- * @returns { double } the number type [since 23 - 23]
- * @returns { int } the number type [since 24]
+ * @returns { double } the number type
  * @syscap SystemCapability.Utils.Lang
  * @stagemodelonly
  * @crossplatform
  * @atomicservice
  * @since 23 dynamic&static
  */
-export type ArrayListComparatorFn<T> = (firstValue: T, secondValue: T) => int;
+export type ArrayListComparatorFn<T> = (firstValue: T, secondValue: T) => double;
 
 /**
  * The type of ArrayList callback function.
